FacebookPostScraping/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
frozenlist==1.5.0
hjson==3.1.0
idna==3.10
jsonpath==0.82.2
jsonpath-ng==1.7.0
multidict==6.1.0
ply==3.11
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
urllib3==2.3.0
wheel==0.44.0
yarl==1.18.3
